requirements.txt
langchain
langchain-community
langchain-google-genai
duckduckgo-search
python-dotenv